+++ tests/matrix/player/gst_rtp.rs
@@ -95,7 +95,8 @@ impl Player for GstRtpPlayer {
         let mut handle_whep = Some(tokio::spawn({
             let ct = ct.clone();
             async move {
-                livetwo::whep::from(ct, output_url, whep_url, None, None, None, None, None).await
+                livetwo::whep::from(ct, output_url, whep_url, None, None, None, None, Vec::new())
+                    .await
             }
         }));